+++ public/scripts/factories/photographer.js
@@ -1,5 +1,5 @@
 function photographerFactory(data) {
-    const { name, portrait } = data;
+    const { name, city, tagline, price, portrait } = data;
 
     const picture = `public/assets/photographers/${portrait}`;
 
@@ -9,13 +9,19 @@ function photographerFactory(data) {
         img.setAttribute("src", picture)
         const h2 = document.createElement( 'h2' );
         h2.textContent = name;
+        const location = document.createElement( 'p' );
+        location.textContent = city;
+        const motto = document.createElement( 'p' );
+        motto.textContent = tagline;
+        const pricing = document.createElement( 'p' );
+        pricing.textContent = price;
         article.appendChild(img);
         article.appendChild(h2);
+        article.appendChild(location);
+        article.appendChild(motto);
+        article.appendChild(pricing);
         return (article);
     }
-    return { name, picture, getUserCardDOM }
+    // Why return getUserCardDOM ?
+    return { name, picture, city, tagline, price, getUserCardDOM }
 }
-
-/* 
-let photographerName = fullData.photographers.name;
-console.log(photographerName); */