+++ CHANGELOG.md
@@ -5,7 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2021-12-01
+### Added
+- Photographers profile pictures in assets folder.
+- Photographers profile picture, location, motto and prince on main page.
+
 ## [0.0.4] - 2021-11-30
+### Added
 - Fetch photographer data and display photographer names on index page.
 
 ## [0.0.3] - 2021-11-25